pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ga4-mcp-server"
version = "1.0.0"
description = "Google Analytics 4 MCP Server for Claude AI"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
requires-python = ">=3.8"
dependencies = [
    "fastmcp>=0.1.0",
    "google-analytics-data>=0.16.0",
]
keywords = ["google-analytics", "mcp", "claude", "ai", "analytics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
Homepage = "https://github.com/yourusername/ga4-mcp-server"
Repository = "https://github.com/yourusername/ga4-mcp-server"
Issues = "https://github.com/yourusername/ga4-mcp-server/issues"

[project.scripts]
ga4-mcp-server = "ga4_mcp_server:main"